Remove only if .vim is link, remove oh-my-zsh submodule

+++ init.sh
@@ -25,7 +25,7 @@ if [ "$right_folder" != "$script_dir" ]; then
 fi
 
 git pull && git submodule update --init
-rm ~/.vim
+[ -h ~/.vim ] && rm ~/.vim
 ln -sf "$(pwd)/vim" ~/.vim
 ln -sf "$(pwd)/vim/vimrc" ~/.vimrc
 ln -sf "$(pwd)/tmux.conf" ~/.tmux.conf
